+++ debian/paasta-tools.links
@@ -30,6 +30,7 @@ opt/venvs/paasta-tools/bin/paasta_list_chronos_jobs usr/bin/list_chronos_jobs
 opt/venvs/paasta-tools/bin/paasta_list_chronos_jobs usr/bin/paasta_list_chronos_jobs
 opt/venvs/paasta-tools/bin/paasta_maintenance.py usr/bin/paasta_maintenance
 opt/venvs/paasta-tools/bin/paasta_metastatus.py usr/bin/paasta_metastatus
+opt/venvs/paasta-tools/bin/paasta_remote_run.py usr/bin/paasta_remote_run
 opt/venvs/paasta-tools/bin/paasta_serviceinit.py usr/bin/paasta_serviceinit
 opt/venvs/paasta-tools/bin/paasta_setup_chronos_job usr/bin/paasta_setup_chronos_job
 opt/venvs/paasta-tools/bin/paasta_setup_chronos_job usr/bin/setup_chronos_job
